+++ sdk/src/wasm/wasm.generated.ts
@@ -375,8 +375,6 @@ export interface Light2D {
     color: Vec4;
     intensity: number;
     radius: number;
-    direction: Vec2;
-    directionZ: number;
     innerAngle: number;
     outerAngle: number;
     shadowSoftness: number;